refactor: move editor layout module
